Disable type-checking for istarmap

+++ src/writers.py
@@ -39,7 +39,7 @@ def write_to_png(screen: np.ndarray, output_file_path: str, progress_bar: bool):
 	width = len(screen)
 	height = len(screen[0])
 
-	screen = screen * 255
+	screen = (screen * 255).astype(int)
 	coords: Iterable = [(x,y) for y in range(height) for x in range(width)]
 	if progress_bar:
 		coords = tqdm(coords)
